pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "consume"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "requests",
    "readability-lxml",
    "anthropic",
    "python-dotenv",
    "selenium",
    "undetected-chromedriver",
    "fpdf2",
    "edge-tts",
]

[project.scripts]
consume = "consume.cli:main"

[project.optional-dependencies]
dev = [
    "pytest",
    "requests-mock",
]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests that require network and API access (deselect with '-m not integration')",
]
